Fix cloud deploy script (#867)
The Python build tool does not handle symlinks. This eliminates the
dependency on the local code mgmt tool from the build script. It also
makes it clear where references to source dirs are actually required.
Once we start pulling `kagglesdk` from pypi.org then this will need to
change. At that time other changes will need to be made to this script
and it should be obvious where the current mod needs to be updated.